requirements.txt
djangorestframework
hypercorn==0.13.2
dynaconf==3.1.7
prometheus-api-client==0.5.1
kubernetes==29.0.0
kubernetes_asyncio==29.0.0
dependency_injector==4.41.0
pytest==7.4.4
pytest-asyncio==0.23.4
async-lru==2.0.4
pyarrow
loguru
django-extensions
ipython
requests-futures
boto3
pyfakefs